Use WXT to build extension
