+++ apps/extension/.gitignore
@@ -1,2 +1,5 @@
-artifacts/
-dist/
+.output
+stats.html
+stats-*.json
+.wxt
+web-ext.config.ts
