+++ apps/extension/README.md
@@ -4,8 +4,8 @@ This is the gw2.me browser extension. Visit https://github.com/GW2Treasures/gw2.
 
 ## Build Instructions
 
-To build the extension you need Node.js 20+ and the latest version of npm.
+This extension uses [WXT](https://wxt.dev/).
 
 1. Install all dependencies by running `npm install`.
-2. Build the extension by running `npm run build`.
-3. You can now find the extension in the `dist/` directory and load it into your browser.
+2. Build the extension by running `npm run build` (`npm run build:firefox` for firefox).
+3. You can now find the extension in the `.output/` directory and load it into your browser.
