+++ apps/extension/eslint.config.mjs
@@ -1,7 +1,8 @@
 import config from '@gw2treasures/eslint-config';
-import { defineConfig, globalIgnores } from 'eslint/config';
+import { defineConfig } from 'eslint/config';
 
 export default defineConfig([
-  globalIgnores(['dist/']),
+  { ignores: ['.wxt', '.output'] },
+  autoImports,
   ...config
 ]);
