--- apps/extension/manifest.json
@@ -1,25 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "gw2.me",
-  "description": "Generate Guild Wars 2 Subtokens",
-  "version": "0.0.6",
-  "permissions": ["identity", "storage"],
-  "action": {
-    "default_popup": "popup.html"
-  },
-  "background": {
-    "service_worker": "background.js",
-    "scripts": ["background.js"]
-  },
-  "icons": {
-    "128": "assets/gw2.me-128.png",
-    "48": "assets/gw2.me-48.png",
-    "32": "assets/gw2.me-32.png",
-    "16": "assets/gw2.me-16.png"
-  },
-  "browser_specific_settings": {
-    "gecko": {
-      "id": "extension@gw2.me"
-    }
-  }
-}
