+++ apps/extension/src/entrypoints/popup/global.css
