--- apps/extension/src/popup.html
@@ -1,6 +1,0 @@
-<html>
-  <body>
-    <div id="root"></div>
-    <script src="./popup.js"></script>
-  </body>
-</html>
