--- apps/extension/src/popup/client.ts
@@ -1,3 +1,0 @@
-import { Gw2MeClient } from '@gw2me/client';
-
-export const client = new Gw2MeClient({ client_id: 'd81b650e-ef90-493a-97a3-871bfcb8063e' });
