+++ apps/extension/src/utils/client.ts
@@ -1,0 +1,3 @@
+import { Gw2MeClient } from '@gw2me/client';
+
+export const client = new Gw2MeClient({ client_id: import.meta.env.WXT_CLIENT_ID });
